Introduced cronjob helper service
