+++ services/121-service/src/cronjob/services/cronjob-execution.service.ts
@@ -1,8 +1,6 @@
 import { Injectable } from '@nestjs/common';
 
-import { CronjobResults } from '@121-service/src/cronjob/interfaces/cronjob-results.interface';
-import { CronjobExecutionMethodName } from '@121-service/src/cronjob/types/cronjob-execution-method-name.type';
-import { env } from '@121-service/src/env';
+import { CronjobExecutionHelperService } from '@121-service/src/cronjob/services/cronjob-execution-helper.service';
 import { ExchangeRatesService } from '@121-service/src/exchange-rates/exchange-rates.service';
 import { IntersolveVoucherService } from '@121-service/src/payments/fsp-integration/intersolve-voucher/intersolve-voucher.service';
 import { IntersolveVoucherCronService } from '@121-service/src/payments/fsp-integration/intersolve-voucher/services/intersolve-voucher-cron.service';
@@ -10,7 +8,6 @@ import { CommercialBankEthiopiaReconciliationService } from '@121-service/src/pa
 import { IntersolveVisaReconciliationService } from '@121-service/src/payments/reconciliation/intersolve-visa-reconciliation/intersolve-visa-reconciliation.service';
 import { IntersolveVoucherReconciliationService } from '@121-service/src/payments/reconciliation/intersolve-voucher-reconciliation/intersolve-voucher-reconciliation.service';
 import { NedbankReconciliationService } from '@121-service/src/payments/reconciliation/nedbank-reconciliation/nedbank-reconciliation.service';
-import { AzureLogService } from '@121-service/src/shared/services/azure-log.service';
 
 @Injectable()
 export class CronjobExecutionService {
@@ -22,17 +19,18 @@ export class CronjobExecutionService {
     private readonly commercialBankEthiopiaReconciliationService: CommercialBankEthiopiaReconciliationService,
     private readonly nedbankReconciliationService: NedbankReconciliationService,
     private readonly exchangeRatesService: ExchangeRatesService,
-    private readonly azureLogService: AzureLogService,
+    private readonly cronjobExecutionHelperService: CronjobExecutionHelperService,
   ) {}
 
   public async cronCancelByRefposIntersolve(): Promise<void> {
-    await this.executeWithLogging('cronCancelByRefposIntersolve', () =>
-      this.intersolveVoucherCronService.cancelByRefposIntersolve(),
+    await this.cronjobExecutionHelperService.executeWithLogging(
+      'cronCancelByRefposIntersolve',
+      () => this.intersolveVoucherCronService.cancelByRefposIntersolve(),
     );
   }
 
   public async cronValidateCommercialBankEthiopiaAccountEnquiries(): Promise<void> {
-    await this.executeWithLogging(
+    await this.cronjobExecutionHelperService.executeWithLogging(
       'cronValidateCommercialBankEthiopiaAccountEnquiries',
       () =>
         this.commercialBankEthiopiaReconciliationService.retrieveAndUpsertAccountEnquiries(),
@@ -40,7 +38,7 @@ export class CronjobExecutionService {
   }
 
   public async cronRetrieveAndUpdatedUnusedIntersolveVouchers(): Promise<void> {
-    await this.executeWithLogging(
+    await this.cronjobExecutionHelperService.executeWithLogging(
       'cronRetrieveAndUpdatedUnusedIntersolveVouchers',
       () =>
         this.intersolveVoucherReconciliationService.cronRetrieveAndUpdatedUnusedIntersolveVouchers(),
@@ -48,88 +46,43 @@ export class CronjobExecutionService {
   }
 
   public async cronRetrieveAndUpdateVisaData(): Promise<void> {
-    await this.executeWithLogging('cronRetrieveAndUpdateVisaData', () =>
-      this.intersolveVisaReconciliationService.retrieveAndUpdateAllWalletsAndCards(),
+    await this.cronjobExecutionHelperService.executeWithLogging(
+      'cronRetrieveAndUpdateVisaData',
+      () =>
+        this.intersolveVisaReconciliationService.retrieveAndUpdateAllWalletsAndCards(),
     );
   }
 
   public async cronSendWhatsappReminders(): Promise<void> {
-    await this.executeWithLogging('cronSendWhatsappReminders', () =>
-      this.intersolveVoucherCronService.sendWhatsappReminders(),
+    await this.cronjobExecutionHelperService.executeWithLogging(
+      'cronSendWhatsappReminders',
+      () => this.intersolveVoucherCronService.sendWhatsappReminders(),
     );
   }
 
   public async cronDoNedbankReconciliation(): Promise<void> {
-    await this.executeWithLogging('cronDoNedbankReconciliation', () =>
-      this.nedbankReconciliationService.doNedbankReconciliation(),
+    await this.cronjobExecutionHelperService.executeWithLogging(
+      'cronDoNedbankReconciliation',
+      () => this.nedbankReconciliationService.doNedbankReconciliation(),
     );
   }
 
   public async cronGetDailyExchangeRates(): Promise<void> {
-    await this.executeWithLogging('cronGetDailyExchangeRates', () =>
-      this.exchangeRatesService.retrieveAndStoreAllExchangeRates(),
+    await this.cronjobExecutionHelperService.executeWithLogging(
+      'cronGetDailyExchangeRates',
+      () => this.exchangeRatesService.retrieveAndStoreAllExchangeRates(),
     );
   }
 
   public async cronRemoveDeprecatedImageCodes(
     mockCurrentDate?: string | undefined,
   ): Promise<number | undefined> {
-    return await this.executeWithLogging('cronRemoveDeprecatedImageCodes', () =>
-      this.intersolveVoucherService.removeDeprecatedImageCodes(mockCurrentDate),
+    return await this.cronjobExecutionHelperService.executeWithLogging(
+      'cronRemoveDeprecatedImageCodes',
+      () =>
+        this.intersolveVoucherService.removeDeprecatedImageCodes(
+          mockCurrentDate,
+        ),
     );
-  }
-
-  private async executeWithLogging(
-    methodName: CronjobExecutionMethodName,
-    fn: () => Promise<number>,
-  ): Promise<number | undefined> {
-    const startMessage = this.createCronjobStartMessage(methodName);
-    this.azureLogService.consoleLogAndTraceAzure(startMessage);
-
-    try {
-      // Execute the cron job function and await its result
-      const batchSize = await fn();
-
-      // Handle the result and log the end message
-      const cronjobResultMessage = this.createCronjobResultMessage({
-        methodName,
-        batchSize,
-        isError: false,
-      });
-      this.azureLogService.consoleLogAndTraceAzure(cronjobResultMessage);
-      return batchSize;
-    } catch (error) {
-      // 1. Log the stack trace to the Node logs
-      console.error(`Error executing cron job ${methodName}:`, error);
-
-      // 2. Log the cronjob end message to Azure Application Insights and trigger an alert
-      const cronjobResultMessage = this.createCronjobResultMessage({
-        methodName,
-        isError: true,
-      });
-      this.azureLogService.logError(new Error(cronjobResultMessage), true);
-    }
-  }
-
-  private createCronjobStartMessage(
-    methodName: CronjobExecutionMethodName,
-  ): string {
-    return `[CRON START] ${env.ENV_NAME} - ${methodName} started at ${new Date().toISOString()}`;
-  }
-
-  private createCronjobResultMessage(cronjobResults: CronjobResults): string {
-    const { methodName, batchSize, isError } = cronjobResults;
-    let message = `[CRON END] ${env.ENV_NAME} - ${methodName} finished at ${new Date().toISOString()}`;
-    const details: string[] = [];
-    if (typeof batchSize === 'number') {
-      details.push(`batchSize=${batchSize}`);
-    }
-    if (isError) {
-      details.push('isError=true');
-    }
-    if (details.length > 0) {
-      message += ` (${details.join(', ')})`;
-    }
-    return message;
   }
 }
